Move the webserviceV1 file to the api/v1 folder
